lei-21-s2-lei-21-s2-1di-g36-43831ee2477f/src/main/java/app/mappers/dto/ParameterCategoryDTO.java
package app.mappers.dto;

import app.domain.model.ParameterCategory;

public class ParameterCategoryDTO {
    private final String code;
    private final String name;

    /**
     * Constructor
     * @param parameterCategory
     */
    public ParameterCategoryDTO(ParameterCategory parameterCategory){
        code = parameterCategory.getCode();
        name = parameterCategory.getName();
    }

    /**
     * This method returns the code of Parameter Category
     * @return
     */
    public String getCode() {
        return code;
    }

    /**
     * This method returns the name of a Parameter Category
     * @return
     */
    public String getname() {
        return name;
    }
}
